Move reportwriter jinja2 filters to another file
