+++ ghostwriter/modules/reportwriter/filters.py
@@ -1,0 +1,254 @@
+"""
+Jinja filters that ghostwriter exposes
+"""
+
+from datetime import timedelta
+import logging
+import re
+
+from bs4 import BeautifulSoup
+from dateutil.parser import parse as parse_datetime
+from dateutil.parser._parser import ParserError
+from django.conf import settings
+from django.utils.dateformat import format as dateformat
+
+from ghostwriter.modules.exceptions import InvalidFilterValue
+
+logger = logging.getLogger(__name__)
+
+
+# Custom Jinja2 filters for DOCX templates
+def filter_severity(findings, allowlist):
+    """
+    Filter a list of findings to return only those with a severity in the allowlist.
+
+    **Parameters**
+
+    ``findings``
+        List of dictionary objects (JSON) for findings
+    ``allowlist``
+        List of strings matching severity categories to allow through filter
+    """
+    filtered_values = []
+    if isinstance(allowlist, list):
+        allowlist = [severity.lower() for severity in allowlist]
+    else:
+        raise InvalidFilterValue(
+            f'Allowlist passed into `filter_severity()` filter is not a list ("{allowlist}"); must be like `["Critical", "High"]`'
+        )
+    try:
+        for finding in findings:
+            if finding["severity"].lower() in allowlist:
+                filtered_values.append(finding)
+    except (KeyError, TypeError):
+        logger.exception("Error parsing ``findings`` as a list of dictionaries: %s", findings)
+        raise InvalidFilterValue(
+            "Invalid list of findings passed into `filter_severity()` filter; must be the `{{ findings }}` object"
+        )
+    return filtered_values
+
+
+def filter_type(findings, allowlist):
+    """
+    Filter a list of findings to return only those with a type in the allowlist.
+
+    **Parameters**
+
+    ``findings``
+        List of dictionary objects (JSON) for findings
+    ``allowlist``
+        List of strings matching severity categories to allow through filter
+    """
+    filtered_values = []
+    if isinstance(allowlist, list):
+        allowlist = [t.lower() for t in allowlist]
+    else:
+        raise InvalidFilterValue(
+            f'Allowlist passed into `filter_type()` filter is not a list ("{allowlist}"); must be like `["Network", "Web"]`'
+        )
+    try:
+        for finding in findings:
+            if finding["finding_type"].lower() in allowlist:
+                filtered_values.append(finding)
+    except (KeyError, TypeError):
+        logger.exception("Error parsing ``findings`` as a list of dictionaries: %s", findings)
+        raise InvalidFilterValue(
+            "Invalid list of findings passed into `filter_type()` filter; must be the `{{ findings }}` object"
+        )
+    return filtered_values
+
+
+def strip_html(s):
+    """
+    Strip HTML tags from the provided HTML while preserving newlines created by
+    ``<br />`` and ``<p>`` tags and spaces.
+
+    **Parameters**
+
+    ``s``
+        String of HTML text to strip down
+    """
+    html = BeautifulSoup(s, "lxml")
+    output = ""
+    for tag in html.descendants:
+        if isinstance(tag, str):
+            output += tag
+        elif tag.name in ("br", "p"):
+            output += "\n"
+    return output.strip()
+
+
+def compromised(targets):
+    """
+    Filter a list of targets to return only those marked as compromised.
+
+    **Parameters**
+
+    ``targets``
+        List of dictionary objects (JSON) for targets
+    """
+    filtered_targets = []
+    try:
+        for target in targets:
+            if target["compromised"]:
+                filtered_targets.append(target)
+    except (KeyError, TypeError):
+        logger.exception("Error parsing ``targets`` as a list of dictionaries: %s", targets)
+        raise InvalidFilterValue(
+            "Invalid list of targets passed into `compromised()` filter; must be the `{{ targets }}` object"
+        )
+    return filtered_targets
+
+
+def add_days(date, days):
+    """
+    Add a number of business days to a date.
+
+    **Parameters**
+
+    ``date``
+        Date string to add business days to
+    ``days``
+        Number of business days to add to the date
+    """
+    new_date = None
+    try:
+        days = int(days)
+    except ValueError:
+        logger.exception("Error parsing ``days`` as an integer: %s", days)
+        raise InvalidFilterValue(f'Invalid integer ("{days}") passed into the `add_days()` filter')
+
+    try:
+        date_obj = parse_datetime(date)
+        # Loop until all days added
+        if days > 0:
+            while days > 0:
+                # Add one day to the date
+                date_obj += timedelta(days=1)
+                # Check if the day is a business day
+                weekday = date_obj.weekday()
+                if weekday >= 5:
+                    # Return to the top (Sunday is 6)
+                    continue
+                # Decrement the number of days to add
+                days -= 1
+        else:
+            # Same as above but in reverse for negative days
+            while days < 0:
+                date_obj -= timedelta(days=1)
+                weekday = date_obj.weekday()
+                if weekday >= 5:
+                    continue
+                days += 1
+        new_date = dateformat(date_obj, settings.DATE_FORMAT)
+    except ParserError:
+        logger.exception("Error parsing ``date`` as a date: %s", date)
+        raise InvalidFilterValue(f'Invalid date string ("{date}") passed into the `add_days()` filter')
+    return new_date
+
+
+def format_datetime(date, new_format):
+    """
+    Change the format of a given date string.
+
+    **Parameters**
+
+    ``date``
+        Date string to modify
+    ``format_str``
+        The format of the provided date
+    """
+    try:
+        date_obj = parse_datetime(date)
+        formatted_date = dateformat(date_obj, new_format)
+    except ParserError:
+        logger.exception("Error parsing ``date`` as a date: %s", date)
+        raise InvalidFilterValue(f'Invalid date string ("{date}") passed into the `format_datetime()` filter')
+    return formatted_date
+
+
+def get_item(lst, index):
+    """
+    Get the item at the specified index in a list.
+
+    **Parameters**
+
+    ``list``
+        List to get item from
+    ``index``
+        Index of item to get
+    """
+    try:
+        return lst[index]
+    except TypeError:
+        logger.exception("Error getting list index %s from this list: %s", index, lst)
+        raise InvalidFilterValue("Invalid list or string passed into the `get_item()` filter")
+    except IndexError:
+        logger.exception("Error getting index %s from this list: %s", index, lst)
+        raise InvalidFilterValue("Invalid or unavailable index passed into the `get_item()` filter")
+
+
+def regex_search(text, regex):
+    """
+    Perform a regex search on the provided text and return the first match.
+
+    **Parameters**
+
+    ``regex``
+        Regular expression to search with
+    ``text``
+        Text to search
+    """
+    match = re.search(regex, text)
+    if match:
+        return match.group(0)
+    return None
+
+
+def filter_tags(objects, allowlist):
+    """
+    Filter a list of objects to return only those with a tag in the allowlist.
+
+    **Parameters**
+
+    ``objects``
+        List of dictionary objects (JSON) for findings
+    ``allowlist``
+        List of strings matching severity categories to allow through filter
+    """
+    filtered_values = []
+    if not isinstance(allowlist, list):
+        raise InvalidFilterValue(
+            f'Allowlist passed into `filter_tags()` filter is not a list ("{allowlist}"); must be like `["xss", "T1651"]`'
+        )
+    try:
+        for obj in objects:
+            common_tags = set(obj["tags"]) & set(allowlist)
+            if common_tags:
+                filtered_values.append(obj)
+    except (KeyError, TypeError):
+        logger.exception("Error parsing object as a list of dictionaries: %s", object)
+        raise InvalidFilterValue(
+            "Invalid list of objects passed into `filter_tags()` filter; must be an object with a `tags` key"
+        )
+    return filtered_values
